requirements.txt
absl-py==1.0.0
APScheduler==3.6.3
astunparse==1.6.3
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
cycler==0.11.0
Flask==2.0.1
flatbuffers==2.0
fonttools==4.32.0
gast==0.5.3
google-api-core==1.28.0
google-api-python-client==2.5.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.44.0
gspread==3.7.0
h5py==3.6.0
httplib2==0.19.1
idna==2.10
importlib-metadata==4.11.3
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
numpy==1.20.3
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
Pillow==9.1.0
pipdeptree==2.0.0
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygsheets==2.0.5
pyparsing==2.4.7
python-dateutil==2.8.1
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tornado==6.1
typing_extensions==4.2.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.4
Werkzeug==2.0.1
wrapt==1.14.0
zipp==3.8.0
